Рефакторинг. Доработан класс захвата изображения. Добавлены классы для работы с категориями.

+++ lib/app_config.py
@@ -3,7 +3,7 @@ import os
 
 class AppConfig:
 
-    def __init__(self, accepted_classes: list, accepted_score: int, capture_source: int, default_email: str):
+    def __init__(self, accepted_classes: list, accepted_score: float, capture_source: int, default_email: str):
         self.__accepted_classes = accepted_classes
         self.__accepted_score = accepted_score
         self.__capture_source = capture_source
@@ -28,7 +28,7 @@ class AppConfig:
     @staticmethod
     def from_env():
         ac = str(os.getenv('accepted_classes')).replace(' ', '').split(',')
-        asc = int(os.getenv('accepted_score'))
+        asc = float(os.getenv('accepted_score'))
         cs = int(os.getenv('capture_source'))
         de = os.getenv('default_email')
